+++ src/components/layout/FNotes.js
@@ -19,8 +19,8 @@ Displays speaker / teacher notes.
     width="33vw"
     style="
       position: absolute;
-      z-index: 100000;
-      right: 90px;
+      z-index: 10000;
+      right: 95px;
       top: 12px;
     "
   >