+++ src/components/layout/FSidebar.js
@@ -25,7 +25,11 @@ Sidebars load can be either inline or load from a file.
     <span>
       <span @click.prevent="open = !open">
         <slot name="button">
-          <a style="color: var(--blue); border-bottom: 1px dotted var(--blue); cursor: alias;">{{ title }}</a>
+          <a style="
+            color: var(--blue);
+            border-bottom: 1px dotted var(--blue);
+            cursor: alias;"
+          >{{ title }}</a>
         </slot>
       </span>
       <f-fade
@@ -37,15 +41,13 @@ Sidebars load can be either inline or load from a file.
           bottom: 0px;
           background: var(--white);
           overflowY: auto;
-          zIndex: 100000000 !important;
+          z-index: 100000000 !important;
         "
         :style="{
           width: width,
-          boxShadow: (orientation == 'right' ? '-5px' : '5px') + ' 0 10px rgba(0,0,0,0.25)',
+          boxShadow: (orientation == 'right' ? '-5px' : '5px') + ' 0 10px rgba(0,0,0,0.15)',
           right: orientation == 'right' ? 0 : '',
           left: orientation == 'left' ? 0 : '',
-          borderLeft: orientation == 'right' ? 'var(--border-width) solid var(--primary)' : '',
-          borderRight: orientation == 'left' ? 'var(--border-width) solid var(--primary)' : '',
         }"
       >
         <a